requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0
redis==5.0.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Logging
structlog==23.2.0

# Error Monitoring (GlitchTip/Sentry)
sentry-sdk[fastapi]==1.40.4

# Rate Limiting
slowapi==0.1.9

# OAuth and Email dependencies
authlib==1.2.1
httpx==0.25.2
cryptography==41.0.7
emails==0.6.0
jinja2==3.1.2
PyJWT==2.8.0

# Testing dependencies
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# SSL compatibility
urllib3>=1.26.0,<3.0.0

# System monitoring (for health checks)
psutil==5.9.8 